Broke the file part of nautilus-directory.[ch] into nautilus-file.[ch].
	* components/music/nautilus-music-view.c,
	components/notes/ntl-notes.c, libnautilus/Makefile.am,
	libnautilus/nautilus-directory-private.h,
	libnautilus/nautilus-directory.h,
	libnautilus/nautilus-directory.c,
	libnautilus/nautilus-file-private.h, libnautilus/nautilus-file.h,
	libnautilus/nautilus-file.c, libnautilus/nautilus-icon-factory.h,
	src/ntl-index-panel.c, src/file-manager/fm-directory-view.h:
	Broke the file part of nautilus-directory.[ch] into
	nautilus-file.[ch].

	* libnautilus/nautilus-string.c (nautilus_strstrip): Added more
	tests. Changed algorithm on strstrip so it's linear instead of
	n^2.

	* libnautilus/gnome-icon-container.c
	(gnome_icon_container_initialize): Fixed font for "larger" zoom
	level. Also made it assert if the font can't be loaded.

	* components/help/hyperbola-nav-index.c (despace):
	Fixed warning.

	* libnautilus/gnome-icon-container-dnd.c,
	libnautilus/nautilus-string.h: Whitespace.

	* src/file-manager/Makefile.am,
	src/file-manager/fm-directory-view.c,
	src/file-manager/fm-properties-window.h,
	src/file-manager/fm-properties-window.c: Added properties window
	hook, but no window yet. The menu item still needs to be added to
	the main menus as well as the context menus.

	* src/file-manager/fm-directory-view.c,
	src/file-manager/fm-directory-view-icons.c,
	src/file-manager/fm-icon-text-window.h,
	src/file-manager/fm-icon-text-window.c: A little cleanup. Shorter
	names.

+++ components/help/hyperbola-nav-index.c
@@ -366,7 +366,7 @@ despace(GString *s)
 	      ctmp = ctmp_s + 2;
 	      i--;
 	      if(i != strlen(ctmp))
-		g_error("i (%d) != strlen(ctmp) (%ld)", i, strlen(ctmp));
+		g_error("i (%d) != strlen(ctmp) (%ld)", i, (long)strlen(ctmp));
 	    }
 	  ctmp_s = NULL;
 	}
